Organize property test in a more reusable manner
